[feat] 배경 이미지 추가
