+++ app/src/main/res/values/strings.xml
@@ -26,4 +26,7 @@
 
     <!-- Resume ResumeListItem -->
     <string name="resumelistitem_period">%s ~ %s</string>
+
+    <!-- ResumeCertificationCard -->
+    <string name="resume_certification_card_small">%s년 %s월 %s일에 획득</string>
 </resources>
